requirements.txt
Django
djangorestframework
pandas
numpy
matplotlib
seaborn